*Applying new model for recognition, settings changed
Until now, Our app couldn't apply new model that user just created. So we had to restart the app for that.
Also, when we change settings(specifically getsture recognition period setting) we also had to restart the app. or reconnect bluetooth.

And this is for applying those changes directly without restarting app. So user can simply make new gesture recognition model or change settings and it will be applied right after that.
How convenience!!