*Starting activity when screen is off + Service stopping..
Important!
1.Now we can start main activity when screen is off.

We had to use gesture when screen is on especially when we are using myo life hacker with activities.

And that means it was not really fully remote controlled application because if screen turned off after timeout... and what? Can't use except music!

So now if user do little finger gesture, it will automatically start main menu activity so user can perform unlock, select menu ...etc

2. Can stop myo life hacker service through the notification bar by clicking "stop"